refactor(schedule): move the watcher lifetime into signal
